requirements.txt
fastapi==0.115.8
kafka-python==2.0.2
SQLAlchemy==2.0.38
starlette==0.45.3
uvicorn==0.34.0
